+++ mongo_datatables/__init__.py
@@ -1,7 +1,7 @@
 """
 MongoDB integration with jQuery DataTables
 """
-__version__ = '1.36.1'
+__version__ = '1.38.0'
 
 from mongo_datatables.datatables import DataTables, DataField
 from mongo_datatables.editor import Editor